third_party/WebKit/LayoutTests/imported/web-platform-tests/html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-invoke-load-expected.txt

This is a testharness.js-based test.
FAIL invoking resource selection with load() assert_equals: networkState after v.load() expected 3 but got 0
Harness: the test ran to completion.
